% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LUADrisk.R
\name{LUADrisk}
\alias{LUADrisk}
\title{Evaluate Lung Adenocarcinoma Survival Risk}
\usage{
LUADrisk(genes_expr, parallel = 1)
}
\arguments{
\item{genes_expr}{Gene expression matrix. Rownames of the input matrix are genes, and colnames of the input matrix are sample names}

\item{parallel}{Set calculation parallel numbers}
}
\value{
Data frame includes signature score, survival risk and gene expression.
}
\description{
Evaluate Lung Adenocarcinoma Survival Risk
}
\examples{
library(LUADrisk)
Sample_risk <- LUADrisk(genes_expr)
}
